Intellij is now happy with the GAE Java8. I can now build and debug locally.
Had to download the whole google cloud system, and fight with it a bunch.

Created a new project off on the side using the latest maven archetype,
 and then migrated those setting into here.

Threw thew JAX-RS stuff under the bus / away, cause I could not get it
 to work.   The breakthru came when I switched over to a BOG standard
 servlet, and the system complained about not being able to find Jolt.

It seems like it wasn't including dependent Jars in the WEB-INF/lib dir.
Fixed that somehow, and now it is reliably working.

In theory could bring the Jax-RS stuff back, but eh, this app is simple
 enough it does not need the complexity.

Also made a favicon.ico so I can stop seeing errors about that in the log.

Notable changes
 - "Deployment" version of the app, aka 7,8,9 is now no longer in the
   .xml files.  It is set when you "deploy".
 - It is more Heroku like now in that, from my box I can push and tell it
   migrate to the version I just pushed.   Neat, but surprised me and
   generally seems dangerous, because on my Intellij build combo box
   run/deploy locally are just one item away from deploy to cloud.
 - Getting rid of the JAX-RS stuff dropped the war file size from 6megs
   to 2megs.
 - It looks like it is "under the covers" making a docker image and then
   pushing that.  It complained about the appversion being in the "app.yaml"
   which is not in the project, but is part of their docker system.